chore(test): use file_test_runner
